Restructure api module
