+++ src/dbus_api/api/manager_2_0/methods.rs
@@ -1,0 +1,148 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+use std::{path::Path, vec::Vec};
+
+use dbus::{
+    self,
+    arg::Array,
+    tree::{MTFn, MethodInfo, MethodResult},
+    Message,
+};
+
+use crate::{
+    dbus_api::{
+        blockdev::create_dbus_blockdev,
+        pool::create_dbus_pool,
+        types::TData,
+        util::{
+            engine_to_dbus_err_tuple, get_next_arg, msg_code_ok, msg_string_ok, tuple_to_option,
+        },
+    },
+    engine::{CreateAction, DeleteAction, PoolUuid},
+};
+
+pub fn create_pool(m: &MethodInfo<MTFn<TData>, TData>) -> MethodResult {
+    let message: &Message = m.msg;
+    let mut iter = message.iter_init();
+
+    let name: &str = get_next_arg(&mut iter, 0)?;
+    let redundancy: (bool, u16) = get_next_arg(&mut iter, 1)?;
+    let devs: Array<&str, _> = get_next_arg(&mut iter, 2)?;
+
+    let blockdevs = devs.map(|x| Path::new(x)).collect::<Vec<&Path>>();
+
+    let object_path = m.path.get_name();
+    let dbus_context = m.tree.get_data();
+    let mut engine = dbus_context.engine.borrow_mut();
+    let result = engine.create_pool(name, &blockdevs, tuple_to_option(redundancy));
+
+    let return_message = message.method_return();
+
+    let default_return: (bool, (dbus::Path<'static>, Vec<dbus::Path<'static>>)) =
+        (false, (dbus::Path::default(), Vec::new()));
+
+    let msg = match result {
+        Ok(pool_uuid_action) => {
+            let results = match pool_uuid_action {
+                CreateAction::Created(uuid) => {
+                    let (_, pool) = get_mut_pool!(engine; uuid; default_return; return_message);
+
+                    let pool_object_path: dbus::Path =
+                        create_dbus_pool(dbus_context, object_path.clone(), uuid, pool);
+
+                    let bd_paths = pool
+                        .blockdevs_mut()
+                        .into_iter()
+                        .map(|(uuid, bd)| {
+                            create_dbus_blockdev(dbus_context, pool_object_path.clone(), uuid, bd)
+                        })
+                        .collect::<Vec<_>>();
+                    (true, (pool_object_path, bd_paths))
+                }
+                CreateAction::Identity => default_return,
+            };
+            return_message.append3(results, msg_code_ok(), msg_string_ok())
+        }
+        Err(x) => {
+            let (rc, rs) = engine_to_dbus_err_tuple(&x);
+            return_message.append3(default_return, rc, rs)
+        }
+    };
+    Ok(vec![msg])
+}
+
+pub fn destroy_pool(m: &MethodInfo<MTFn<TData>, TData>) -> MethodResult {
+    let message: &Message = m.msg;
+    let mut iter = message.iter_init();
+
+    let pool_path: dbus::Path<'static> = get_next_arg(&mut iter, 0)?;
+
+    let dbus_context = m.tree.get_data();
+
+    let default_return = (false, uuid_to_string!(PoolUuid::nil()));
+    let return_message = message.method_return();
+
+    let pool_uuid = match m
+        .tree
+        .get(&pool_path)
+        .and_then(|op| op.get_data().as_ref())
+        .map(|d| d.uuid)
+    {
+        Some(uuid) => uuid,
+        None => {
+            return Ok(vec![return_message.append3(
+                default_return,
+                msg_code_ok(),
+                msg_string_ok(),
+            )]);
+        }
+    };
+
+    let msg = match dbus_context.engine.borrow_mut().destroy_pool(pool_uuid) {
+        Ok(DeleteAction::Deleted(uuid)) => {
+            dbus_context
+                .actions
+                .borrow_mut()
+                .push_remove(&pool_path, m.tree);
+            return_message.append3(
+                (true, uuid_to_string!(uuid)),
+                msg_code_ok(),
+                msg_string_ok(),
+            )
+        }
+        Ok(DeleteAction::Identity) => {
+            return_message.append3(default_return, msg_code_ok(), msg_string_ok())
+        }
+        Err(err) => {
+            let (rc, rs) = engine_to_dbus_err_tuple(&err);
+            return_message.append3(default_return, rc, rs)
+        }
+    };
+    Ok(vec![msg])
+}
+
+pub fn configure_simulator(m: &MethodInfo<MTFn<TData>, TData>) -> MethodResult {
+    let message = m.msg;
+    let mut iter = message.iter_init();
+
+    let denominator: u32 = get_next_arg(&mut iter, 0)?;
+
+    let dbus_context = m.tree.get_data();
+    let result = dbus_context
+        .engine
+        .borrow_mut()
+        .configure_simulator(denominator);
+
+    let return_message = message.method_return();
+
+    let msg = match result {
+        Ok(_) => return_message.append2(msg_code_ok(), msg_string_ok()),
+        Err(err) => {
+            let (rc, rs) = engine_to_dbus_err_tuple(&err);
+            return_message.append2(rc, rs)
+        }
+    };
+    Ok(vec![msg])
+}
